Oauth flow change; moved make request logic to HHAsyncClient; UserService start; defined AbstractRedisRepository; Refined User model

+++ backend/app/application/hh/auth/oauth.py
@@ -1,9 +1,8 @@
 import secrets
-from urllib.parse import urlencode
 
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.application.hh.auth.dto import OauthTokensDTO
+from app.application.hh.auth.dto import HHUserDTO, OauthTokensDTO
 from app.infrastructure.clients import HHAsyncClient
 from app.infrastructure.database import RedisStateRepository
 from app.infrastructure.database.users.repositories import HeadHunterTokenRepository
@@ -11,68 +10,41 @@ from app.infrastructure.database.users.repositories import HeadHunterTokenReposi
 
 class HHOAuthService():
     def __init__(
-            self,
-            client_id: str, secret_key: str,
-            redirect_uri: str,
-            session: AsyncSession,
+            self, session: AsyncSession,
             state_repo: RedisStateRepository,
             hh_token_repo: HeadHunterTokenRepository
         ) -> None:
-        self.client_id = client_id
-        self.secret_key = secret_key
-        self.redirect_uri = redirect_uri
-
-        self.state_repo = state_repo
-        
         self.session = session
+        self.state_repo = state_repo
+
         self.hh_token_repo = hh_token_repo
-        self.hh_authorize_url = 'https://hh.ru/oauth/authorize'
-        self.hh_token_url = 'https://api.hh.ru/token'
         self.hh_client = HHAsyncClient
     
-    async def create_authorize_url(self):
+    async def get_user_authorize_url(self) -> str:
         state = await self.generate_state()
-
-        query = urlencode({
-            "response_type": "code",
-            "client_id": self.client_id,
-            "redirect_uri": self.redirect_uri,
-            "state": state
-        })
-
-        return f"{self.hh_authorize_url}?{query}"
+        return self.hh_client.create_user_authorize_url(state)
     
-    async def get_tokens(self, authorization_code: str, state: str) -> str:
-        await self.validate_state_and_get_user(state)
-        
-        tokens_data = await self.make_token_request(authorization_code)
-        tokens_dto = OauthTokensDTO(**tokens_data)
-        
-        return tokens_dto
+    async def authorize_user(
+            self, authorization_code: str, state: str
+        ) -> tuple[HHUserDTO, OauthTokensDTO]:
+        await self.validate_state(state)
+
+        hh_tokens_dto = await self.get_hh_tokens(authorization_code)
+        hh_user_dto = await self.get_hh_user_info(hh_tokens_dto.access_token)
+
+        return hh_user_dto, hh_tokens_dto
     
-    async def make_token_request(self, authorization_code: str):
-        headers = {
-            "Content-Type": "application/x-www-form-urlencoded"
-        }
-
-        data = {
-            "grant_type": "authorization_code",
-            "client_id": self.client_id,
-            "client_secret": self.secret_key,
-            "redirect_uri": self.redirect_uri,
-            "code": authorization_code,
-        }
-
+    async def get_hh_tokens(self, authorization_code: str) -> OauthTokensDTO:
         async with self.hh_client() as client:
-            token_response = await client.post(
-                self.hh_token_url,
-                headers=headers,
-                data=data,
-            )
-
-        token_data = token_response.json()
-
-        return token_data
+            tokens_data = await client.make_token_request(authorization_code)
+        
+        return OauthTokensDTO(**tokens_data)
+    
+    async def get_hh_user_info(self, access_token: str) -> HHUserDTO:
+        async with self.hh_client() as client:
+            hh_user_data = await client.make_me_request(access_token)
+            
+        return HHUserDTO(**hh_user_data)
 
     async def generate_state(self) -> str:
         state = secrets.token_urlsafe(16)
@@ -80,13 +52,12 @@ class HHOAuthService():
 
         return state
     
-    async def validate_state_and_get_user(self, state: str) -> int:
-        await self.state_repo.get(state)
-        # if not telegram_id:
-            # raise RuntimeError('Invalid or expired state')
+    async def validate_state(self, state: str) -> int:
+        result = await self.state_repo.get(state)
+
+        if not result:
+            raise RuntimeError('CSRF protection failed')  # TODO change to custom app exception
         
         await self.state_repo.delete(state)
-
-        # return int(telegram_id)
